+++ lib/src/core/provider/language_service.dart
@@ -1,0 +1,17 @@
+import 'package:injectable/injectable.dart';
+import 'package:shared_preferences/shared_preferences.dart';
+
+@LazySingleton()
+class LanguageService {
+  static const _languageKey = 'selected_language';
+
+  Future<void> saveSelectedLanguage(String languageCode) async {
+    final prefs = await SharedPreferences.getInstance();
+    await prefs.setString(_languageKey, languageCode);
+  }
+
+  Future<String?> getSelectedLanguage() async {
+    final prefs = await SharedPreferences.getInstance();
+    return prefs.getString(_languageKey);
+  }
+}
